back/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
pymysql
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic
pydantic-settings
python-multipart
